day01/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
idna==3.4
piexif==1.1.3
Pillow==10.0.0
pycparser==2.21
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
